Revert "break into submodules"
This reverts commit a43912e0efd5e7f34686e63d7baf4b8f1a1b99b2.
