+++ pymatviz/ptable.py
@@ -1,40 +1,539 @@
-"""Periodic table plots powered by matplotlib."""
+"""Various periodic table heatmaps with matplotlib and plotly."""
 
 from __future__ import annotations
 
+import itertools
 import math
 import warnings
-from collections.abc import Sequence
-from typing import TYPE_CHECKING, Literal, cast
+from collections.abc import Iterable, Sequence
+from typing import TYPE_CHECKING, Literal, Union, cast, get_args
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import plotly.express as px
+import plotly.figure_factory as ff
 from matplotlib.colors import Colormap, LogNorm, Normalize
 from matplotlib.patches import Rectangle
 from matplotlib.ticker import LogFormatter, LogLocator
 from matplotlib.typing import ColorType
-from pymatgen.core import Element
+from pandas.api.types import is_numeric_dtype, is_string_dtype
+from pymatgen.core import Composition, Element
 
-from pymatviz.colors import ELEM_COLORS_JMOL, ELEM_COLORS_VESTA, ELEM_TYPE_COLORS
-from pymatviz.enums import ElemColorMode, ElemColors, ElemCountMode, Key
-from pymatviz.process_data import count_elements
-from pymatviz.ptable._process_data import (
-    PTableData,
+from pymatviz._preprocess_data import (
     SupportedDataType,
     SupportedValueType,
+    get_df_nest_level,
 )
-from pymatviz.utils import (
-    ElemValues,
-    df_ptable,
-    get_cbar_label_formatter,
-    pick_bw_for_contrast,
-)
+from pymatviz.colors import ELEM_COLORS_JMOL, ELEM_COLORS_VESTA, ELEM_TYPE_COLORS
+from pymatviz.enums import ElemColorMode, ElemColors, ElemCountMode, Key
+from pymatviz.utils import df_ptable, get_cbar_label_formatter, pick_bw_for_contrast
 
 
 if TYPE_CHECKING:
-    from typing import Any, Callable
+    from typing import Any, Callable, Self
+
+    import plotly.graph_objects as go
+    from numpy.typing import NDArray
+
+
+ElemValues = Union[dict[Union[str, int], float], pd.Series, Sequence[str]]
+
+
+def count_elements(
+    values: ElemValues,
+    count_mode: ElemCountMode = ElemCountMode.composition,
+    exclude_elements: Sequence[str] = (),
+    fill_value: float | None = 0,
+) -> pd.Series:
+    """Count element occurrence in list of formula strings or dict-like compositions.
+    If passed values are already a map from element symbol to counts, ensure the
+    data is a pd.Series filled with zero values for missing element symbols.
+
+    Provided as standalone function for external use or to cache long computations.
+    Caching long element counts is done by refactoring
+        ptable_heatmap(long_list_of_formulas) # slow
+    to
+        elem_counts = count_elements(long_list_of_formulas) # slow
+        ptable_heatmap(elem_counts) # fast, only rerun this line to update the plot
+
+    Args:
+        values (dict[str, int | float] | pd.Series | list[str]): Iterable of
+            composition strings/objects or map from element symbols to heatmap values.
+        count_mode ("(element|fractional|reduced)_composition"):
+            Only used when values is a list of composition strings/objects.
+            - composition (default): Count elements in each composition as is,
+                i.e. without reduction or normalization.
+            - fractional_composition: Convert to normalized compositions in which the
+                amounts of each species sum to before counting.
+                Example: Fe2 O3 -> Fe0.4 O0.6
+            - reduced_composition: Convert to reduced compositions (i.e. amounts
+                normalized by greatest common denominator) before counting.
+                Example: Fe4 P4 O16 -> Fe P O4.
+            - occurrence: Count the number of times each element occurs in a list of
+                formulas irrespective of compositions. E.g. [Fe2 O3, Fe O, Fe4 P4 O16]
+                counts to {Fe: 3, O: 3, P: 1}.
+        exclude_elements (Sequence[str]): Elements to exclude from the count. Defaults
+            to ().
+        fill_value (float | None): Value to fill in for missing elements. Defaults to 0.
+
+    Returns:
+        pd.Series: Map element symbols to heatmap values.
+    """
+    valid_count_modes = list(ElemCountMode.key_val_dict())
+    if count_mode not in valid_count_modes:
+        raise ValueError(f"Invalid {count_mode=} must be one of {valid_count_modes}")
+    # Ensure values is Series if we got dict/list/tuple
+    srs = pd.Series(values)
+
+    if is_numeric_dtype(srs):
+        pass
+    elif is_string_dtype(srs) or {*map(type, srs)} <= {str, Composition}:
+        # all items are formula strings or Composition objects
+        if count_mode == "occurrence":
+            srs = pd.Series(
+                itertools.chain.from_iterable(
+                    map(str, Composition(comp, allow_negative=True)) for comp in srs
+                )
+            ).value_counts()
+        else:
+            attr = (
+                "element_composition" if count_mode == Key.composition else count_mode
+            )
+            srs = pd.DataFrame(
+                getattr(Composition(formula, allow_negative=True), attr).as_dict()
+                for formula in srs
+            ).sum()  # sum up element occurrences
+    else:
+        raise ValueError(
+            "Expected values to be map from element symbols to heatmap values or "
+            f"list of compositions (strings or Pymatgen objects), got {values}"
+        )
+
+    try:
+        # If index consists entirely of strings representing integers, convert to ints
+        srs.index = srs.index.astype(int)
+    except (ValueError, TypeError):
+        pass
+
+    if pd.api.types.is_integer_dtype(srs.index):
+        # If index is all integers, assume they represent atomic
+        # numbers and map them to element symbols (H: 1, He: 2, ...)
+        idx_min, idx_max = srs.index.min(), srs.index.max()
+        if idx_max > 118 or idx_min < 1:
+            raise ValueError(
+                "element value keys were found to be integers and assumed to represent "
+                f"atomic numbers, but values range from {idx_min} to {idx_max}, "
+                "expected range [1, 118]."
+            )
+        map_atomic_num_to_elem_symbol = (
+            df_ptable.reset_index().set_index("atomic_number").symbol
+        )
+        srs.index = srs.index.map(map_atomic_num_to_elem_symbol)
+
+    # Ensure all elements are present in returned Series (with value zero if they
+    # weren't in values before)
+    srs = srs.reindex(df_ptable.index, fill_value=fill_value).rename("count")
+
+    if len(exclude_elements) > 0:
+        if isinstance(exclude_elements, str):
+            exclude_elements = [exclude_elements]
+        if isinstance(exclude_elements, tuple):
+            exclude_elements = list(exclude_elements)
+        try:
+            srs = srs.drop(exclude_elements)
+        except KeyError as exc:
+            bad_symbols = ", ".join(x for x in exclude_elements if x not in srs)
+            raise ValueError(
+                f"Unexpected symbol(s) {bad_symbols} in {exclude_elements=}"
+            ) from exc
+
+    return srs
+
+
+class PTableData:
+    """Hold data for ptable plotters.
+
+    Attributes:
+        data (pd.DataFrame): The preprocessed DataFrame with element names
+            as index and values as columns.
+        index_col (str): The index column header.
+        val_col (str): The value column header.
+        anomalies (dict[str, set["nan", "inf"]]): An element to anomalies mapping.
+        nest_level (int): The nest level of DataFrame.
+
+    Example:
+        >>> data_dict: dict = {
+            "H": 1.0,
+            "He": [2.0, 4.0],
+            "Li": [[6.0, 8.0], [10.0, 12.0]],
+        }
+
+        OR
+        >>> data_df: pd.DataFrame = pd.DataFrame(
+            data_dict.items(),
+            columns=["element", "heat_val"]
+            ).set_index("element")
+
+        OR
+        >>> data_series: pd.Series = pd.Series(data_dict)
+
+        >>> ptable_data = PTableData(data_dict / df / series)
+
+        >>> ptable_data.data
+            Element   Value
+            H         [1.0, ]
+            He        [2.0, 4.0]
+            Li        [[6.0, 8.0], [10.0, 11.0]]
+
+        Metadata (ptable_data.data.attrs):
+            vmin: 1.0
+            mean: 6.0
+            vmax: 11.0
+    """
+
+    def __init__(
+        self,
+        data: SupportedDataType,
+        *,
+        index_col: str = Key.element,
+        val_col: str = Key.heat_val,
+        check_missing: bool = True,
+        missing_strategy: Literal["zero", "mean"] = "mean",
+        check_infinity: bool = True,
+        normalize: bool = False,
+    ) -> None:
+        """Preprocess data, including:
+            - Convert all data types to pd.DataFrame.
+            - Replace missing values (NaN) by selected strategy.
+            - Replace infinities with vmax(∞) or vmin(-∞).
+            - Write vmin/mean/vmax as metadata into the DataFrame.
+
+        Args:
+            data (dict[str, float | Sequence[float]] | pd.DataFrame | pd.Series):
+                Input data to preprocess.
+            index_col (str, optional): The index column header. Defaults to Key.index.
+            val_col (str, optional): The value column header. Defaults to Key.heat_val.
+            check_missing (bool): Whether to check and replace missing values.
+            missing_strategy ("zero", "mean"): Missing value replacement strategy.
+                - zero: Replace with zero.
+                - mean: Replace with mean value.
+            check_infinity (bool): Whether to check and replace infinities.
+            normalize (bool): Whether to normalize data.
+        """
+        # Convert and set data, and write metadata
+        self.index_col = index_col
+        self.val_col = val_col
+        self.data = data
+
+        # Get nest level
+        self.nest_level = get_df_nest_level(self.data, col=self.val_col)
+
+        # Replace missing values and infinities, and update anomalies
+        self.anomalies: dict[str, set[Literal["inf", "nan"]]] | Literal["NA"] = (
+            {} if (check_missing or check_infinity) else "NA"
+        )
+        if check_missing:
+            self.check_and_replace_missing(strategy=missing_strategy)
+        if check_infinity:
+            self.check_and_replace_infinity()
+
+        # Normalize data by the total sum
+        if normalize:
+            self.normalize()
+
+    @property
+    def data(self) -> pd.DataFrame:
+        """The internal data as pd.DataFrame."""
+        return self._data
+
+    @data.setter
+    def data(self, data: SupportedDataType) -> None:
+        """Preprocess and set data, including:
+        - Convert supported data types to DataFrame.
+        - Convert all values to NumPy array.
+        - Write vmin/mean/vmax as metadata into the DataFrame.
+        """
+        # Convert supported data types to DataFrame
+        if isinstance(data, pd.DataFrame):
+            data = self._format_pd_dataframe(data)
+
+        elif isinstance(data, pd.Series):
+            data = data.to_frame(name=self.val_col)
+            data.index.name = self.index_col
+
+        elif isinstance(data, dict):
+            data = pd.DataFrame(
+                data.items(), columns=[self.index_col, self.val_col]
+            ).set_index(self.index_col)
+
+        else:
+            type_name = type(data).__name__
+            raise TypeError(
+                f"{type_name} unsupported, choose from {get_args(SupportedDataType)}"
+            )
+
+        # Convert all values to NumPy array
+        data[self.val_col] = data[self.val_col].apply(
+            lambda val: np.array(list(val))
+            if isinstance(val, Iterable) and not isinstance(val, str)
+            else np.array([val])
+        )
+
+        self._data = data
+
+        # Write vmin/mean/vmax as metadata into the DataFrame
+        # Note: this is nested inside the setter so that metadata
+        # would be updated whenever data gets updated
+        self._write_meta_data()
+
+    def _write_meta_data(self) -> None:
+        """Parse meta data and write into data.attrs.
+
+        Currently would handle the following:
+            vmin: The min value.
+            mean: The mean value.
+            vmax: The max value.
+        """
+        numeric_values = pd.to_numeric(
+            self._data[self.val_col].explode().explode().explode(), errors="coerce"
+        )
+        self._data.attrs["vmin"] = numeric_values.min()
+        self._data.attrs["mean"] = numeric_values.mean()
+        self._data.attrs["vmax"] = numeric_values.max()
+
+    @staticmethod
+    def _format_pd_dataframe(df: pd.DataFrame) -> pd.DataFrame:
+        """Format pd.DataFrame that may not meet expected format."""
+
+        def fix_df_elem_as_col(df: pd.DataFrame) -> pd.DataFrame:
+            """Fix pd.DataFrame where elements are in a single column."""
+            # Copy and reset index to move element names to a column
+            new_df = df.copy().reset_index()
+
+            # Find the column with element names
+            elem_col_name = None
+            for col in new_df.columns:
+                if set(new_df[col]).issubset(set(map(str, Element))):
+                    elem_col_name = col
+                    break
+
+            # Fix failed: cannot find the elements column
+            if elem_col_name is None:
+                return None
+
+            # Rename the column with elements and set it as index
+            new_df = new_df.rename(columns={elem_col_name: Key.element})
+            new_df = new_df.set_index(Key.element)
+
+            # Zip the remaining values into a single column
+            value_cols = [
+                col_name for col_name in new_df.columns if col_name != Key.element
+            ]
+            new_df[Key.heat_val] = new_df[value_cols].apply(list, axis="columns")
+
+            # Drop the old value columns
+            return new_df[[Key.heat_val]]
+
+        # Check if input DataFrame is in expected format
+        if Key.element == df.index.name and Key.heat_val in df.columns:
+            return df
+
+        # Re-format it to expected
+        warnings.warn("DataFrame has unexpected format, trying to fix.", stacklevel=2)
+
+        # Try to search for elements as a column
+        fixed_df = fix_df_elem_as_col(df)
+        if fixed_df is not None:
+            return fixed_df
+
+        # Try to search for elements as a row
+        fixed_df = fix_df_elem_as_col(df.transpose())
+        if fixed_df is not None:
+            return fixed_df
+
+        raise KeyError(
+            f"Cannot handle dataframe={df}, needs row/column named "
+            f"{Key.element} and {Key.heat_val}"
+        )
+
+    @classmethod
+    def from_formulas(
+        cls,
+        formulas: Sequence[str] | dict[str, float],
+        count_mode: ElemCountMode = ElemCountMode.composition,
+        exclude_elements: Sequence[str] = (),
+    ) -> Self:
+        """Initialize PTableData from sequences of chemical formulas
+        or dict.
+
+        TODO: migrate the code instead of wrapping.
+
+        TODO: migrate unit test.
+
+        Args:
+            formulas (dict[str, float] | pd.Series | list[str]): Sequence of
+                composition strings/objects or map from element symbols to values.
+            count_mode: Only for formulas as a sequence of composition strings/objects.
+                - "composition" (default): Count elements in each composition as is,
+                    i.e. without reduction or normalization.
+                - "fractional_composition": Convert to normalized compositions in
+                    which the amounts of each species sum to before counting.
+                    Example: "Fe2 O3" -> {Fe: 0.4, O: 0.6}
+                - "reduced_composition": Convert to reduced compositions (i.e. amounts
+                    normalized by greatest common denominator) before counting.
+                    Example: "Fe4 P4 O16" -> {Fe: 1, P: 1 O: 4}.
+                - "occurrence": Count the number of times each element occurs
+                    irrespective of compositions. E.g. ["Fe2 O3", "Fe O", "Fe4 P4 O16"]
+                    counts to {Fe: 3, O: 3, P: 1}.
+            exclude_elements (Sequence[str]): Elements to exclude.
+
+        """
+        # Convert sequences of chemical formulas
+        data = count_elements(
+            values=formulas, count_mode=count_mode, exclude_elements=exclude_elements
+        )
+
+        return cls(data)
+
+    def normalize(self) -> None:
+        """Normalize data by the total sum."""
+        total_sum = self._data.map(np.sum).sum().sum()
+        self.apply(lambda x: x / total_sum)
+
+    def apply(self, func: Callable[[Any], Any]) -> None:
+        """Apply a function to all values in value column.
+
+        Args:
+            func (Callable): The function to be applied.
+        """
+        original_data = self.data
+        original_data[self.val_col] = original_data[self.val_col].apply(func)
+
+        # Ensure metadata is updated by using the setter method
+        self.data = original_data
+
+    def drop_elements(self, elements: Sequence[str]) -> None:
+        """Drop selected elements from data.
+
+        Args:
+            elements (Sequence[str]): Elements to drop.
+        """
+        if elements:
+            original_data = self.data
+            df_dropped = original_data.drop(elements, axis=0)
+
+            self.data = df_dropped
+
+    def check_and_replace_missing(
+        self,
+        strategy: Literal["zero", "mean"],
+    ) -> bool:
+        """Check if there is missing value (NaN) in the values column,
+        and replace them according to selected strategy.
+
+        Would update the anomalies attribute.
+
+        Args:
+            strategy ("zero", "mean"): Missing value replacement strategy.
+                - zero: Replace with zero.
+                - mean: Replace with mean value.
+
+        Returns:
+            bool: Whether the column contains NaN.
+        """
+        # Convert to numeric, forcing non-numeric types to NaN
+        all_values = self.data[self.val_col].explode().explode().explode()
+        all_values = pd.to_numeric(all_values, errors="coerce")
+
+        has_nan = False
+        self.anomalies = cast(dict[str, set[Literal["inf", "nan"]]], self.anomalies)
+
+        for elem, value in all_values.items():
+            if np.isnan(value):
+                if elem not in self.anomalies:
+                    self.anomalies[elem] = set()
+                self.anomalies[elem].add("nan")
+                has_nan = True
+
+        if has_nan:
+            if self.nest_level > 1:
+                raise NotImplementedError(
+                    "Unable to replace NaN and inf for nest_level>1"
+                )
+
+            warnings.warn("NaN found in data", stacklevel=2)
+
+            # Generate and apply replacement
+            def replace_nan(val: NDArray | float) -> NDArray:
+                """Replace missing value based on selected strategy.
+
+                Args:
+                    val (NDarray | float): Value to be processed.
+                """
+                replace_val = (
+                    0 if strategy == "zero" else all_values[all_values != np.inf].mean()
+                )
+
+                if isinstance(val, np.ndarray):
+                    return np.array([replace_val if np.isnan(v) else v for v in val])
+
+                return replace_val if np.isnan(val) else val
+
+            self.apply(replace_nan)
+
+        return has_nan
+
+    def check_and_replace_infinity(self) -> bool:
+        """Check if there is infinity in the data column, and
+        replace them with vmax(∞) or vmin(-∞) if any.
+
+        Would update the anomalies attribute.
+
+        Returns:
+            bool: Whether the column contains NaN.
+        """
+        # Convert to numeric, forcing non-numeric types to NaN
+        all_values = self.data[self.val_col].explode().explode().explode()
+        all_values = pd.to_numeric(all_values, errors="coerce")
+
+        has_inf = False
+        self.anomalies = cast(dict[str, set[Literal["inf", "nan"]]], self.anomalies)
+
+        for elem, value in all_values.items():
+            if np.isinf(value):
+                if elem not in self.anomalies:
+                    self.anomalies[elem] = set()
+                self.anomalies[elem].add("inf")
+                has_inf = True
+
+        if has_inf:
+            if self.nest_level > 1:
+                raise NotImplementedError(
+                    "Unable to replace NaN and inf for nest_level>1."
+                )
+
+            warnings.warn("Infinity found in data", stacklevel=2)
+
+            # Generate and apply replacement
+            def replace_inf(val: NDArray | float) -> NDArray:
+                """Replace infinities."""
+                replacements = {
+                    np.inf: all_values[all_values != np.inf].max(),
+                    -np.inf: all_values[all_values != -np.inf].min(),
+                }
+
+                if isinstance(val, np.ndarray):
+                    return np.array(
+                        [replacements.get(v, v) if np.isinf(v) else v for v in val]
+                    )
+
+                return replacements[val] if np.isinf(val) else val
+
+            self.apply(replace_inf)
+
+        return has_inf
 
 
 class PTableProjector:
@@ -1319,6 +1818,265 @@ def ptable_heatmap_ratio(
     #     bbox = dict(facecolor=color, edgecolor="gray")
     #     fig.text(0.005, y_pos, txt, fontsize=10, bbox=bbox, transform=fig.transAxes)
 
+    return fig
+
+
+def ptable_heatmap_plotly(
+    values: ElemValues,
+    *,
+    count_mode: ElemCountMode = ElemCountMode.composition,
+    colorscale: str | Sequence[str] | Sequence[tuple[float, str]] = "viridis",
+    show_scale: bool = True,
+    show_values: bool = True,
+    heat_mode: Literal["value", "fraction", "percent"] | None = "value",
+    fmt: str | None = None,
+    hover_props: Sequence[str] | dict[str, str] | None = None,
+    hover_data: dict[str, str | int | float] | pd.Series | None = None,
+    font_colors: Sequence[str] = (),
+    gap: float = 5,
+    font_size: int | None = None,
+    bg_color: str | None = None,
+    color_bar: dict[str, Any] | None = None,
+    cscale_range: tuple[float | None, float | None] = (None, None),
+    exclude_elements: Sequence[str] = (),
+    log: bool = False,
+    fill_value: float | None = None,
+    label_map: dict[str, str] | Callable[[str], str] | Literal[False] | None = None,
+    **kwargs: Any,
+) -> go.Figure:
+    """Create a Plotly figure with an interactive heatmap of the periodic table.
+    Supports hover tooltips with custom data or atomic reference data like
+    electronegativity, atomic_radius, etc. See kwargs hover_data and hover_props, resp.
+
+    Args:
+        values (dict[str, int | float] | pd.Series | list[str]): Map from element
+            symbols to heatmap values e.g. dict(Fe=2, O=3) or iterable of composition
+            strings or Pymatgen composition objects.
+        count_mode ("composition" | "fractional_composition" | "reduced_composition"):
+            Reduce or normalize compositions before counting. See count_elements() for
+            details. Only used when values is list of composition strings/objects.
+        colorscale (str | list[str] | list[tuple[float, str]]): Color scale for heatmap.
+            Defaults to "viridis". See plotly.com/python/builtin-colorscales for names
+            of other builtin color scales. Note "YlGn" and px.colors.sequential.YlGn are
+            equivalent. Custom scales are specified as ["blue", "red"] or
+            [[0, "rgb(0,0,255)"], [0.5, "rgb(0,255,0)"], [1, "rgb(255,0,0)"]].
+        show_scale (bool): Whether to show a bar for the color scale. Defaults to True.
+        show_values (bool): Whether to show numbers on heatmap tiles. Defaults to True.
+        heat_mode ("value" | "fraction" | "percent" | None): Whether to display heat
+            values as is (value), normalized as a fraction of the total, as percentages
+            or not at all (None). Defaults to "value".
+            "fraction" and "percent" can be used to make the colors in different
+            periodic table heatmap plots comparable.
+        fmt (str): f-string format option for heat labels. Defaults to ".1%"
+            (1 decimal place) if heat_mode="percent" else ".3g".
+        hover_props (list[str] | dict[str, str]): Elemental properties to display in the
+            hover tooltip. Can be a list of property names to display only the values
+            themselves or a dict mapping names to what they should display as. E.g.
+            dict(atomic_mass="atomic weight") will display as `"atomic weight = {x}"`.
+            Defaults to None.
+            Available properties are: symbol, row, column, name,
+            atomic_number, atomic_mass, n_neutrons, n_protons, n_electrons, period,
+            group, phase, radioactive, natural, metal, nonmetal, metalloid, type,
+            atomic_radius, electronegativity, first_ionization, density, melting_point,
+            boiling_point, number_of_isotopes, discoverer, year, specific_heat,
+            n_shells, n_valence.
+        hover_data (dict[str, str | int | float] | pd.Series): Map from element symbols
+            to additional data to display in the hover tooltip. dict(Fe="this appears in
+            the hover tooltip on a new line below the element name"). Defaults to None.
+        font_colors (list[str]): One color name or two for [min_color, max_color].
+            min_color is applied to annotations with heatmap values less than
+            (max_val - min_val) / 2. Defaults to None, meaning auto-set to maximize
+            contrast with color scale: white text for dark background and vice versa.
+            swapped depending on the colorscale.
+        gap (float): Gap in pixels between tiles of the periodic table. Defaults to 5.
+        font_size (int): Element symbol and heat label text size. Any valid CSS size
+            allowed. Defaults to automatic font size based on plot size. Element symbols
+            will be bold and 1.5x this size.
+        bg_color (str): Plot background color. Defaults to "rgba(0, 0, 0, 0)".
+        color_bar (dict[str, Any]): Plotly colorbar properties documented at
+            https://plotly.com/python/reference#heatmap-colorbar. Defaults to
+            dict(orientation="h"). Commonly used keys are:
+            - title: colorbar title
+            - titleside: "top" | "bottom" | "right" | "left"
+            - tickmode: "array" | "auto" | "linear" | "log" | "date" | "category"
+            - tickvals: list of tick values
+            - ticktext: list of tick labels
+            - tickformat: f-string format option for tick labels
+            - len: fraction of plot height or width depending on orientation
+            - thickness: fraction of plot height or width depending on orientation
+        cscale_range (tuple[float | None, float | None]): Colorbar range. Defaults to
+            (None, None) meaning the range is automatically determined from the data.
+        exclude_elements (list[str]): Elements to exclude from the heatmap. E.g. if
+            oxygen overpowers everything, you can do exclude_elements=["O"].
+            Defaults to ().
+        log (bool): Whether to use a logarithmic color scale. Defaults to False.
+            Piece of advice: colorscale="viridis" and log=True go well together.
+        fill_value (float | None): Value to fill in for missing elements. Defaults to 0.
+        label_map (dict[str, str] | Callable[[str], str] | None): Map heat values (after
+            string formatting) to target strings. Set to False to disable. Defaults to
+            dict.fromkeys((np.nan, None, "nan"), " ") so as not to display "nan" for
+            missing values.
+        **kwargs: Additional keyword arguments passed to
+            plotly.figure_factory.create_annotated_heatmap().
+
+    Returns:
+        Figure: Plotly Figure object.
+    """
+    if log and heat_mode in ("fraction", "percent"):
+        raise ValueError(
+            "Combining log color scale and heat_mode='fraction'/'percent' unsupported"
+        )
+    if len(cscale_range) != 2:
+        raise ValueError(f"{cscale_range=} should have length 2")
+
+    if isinstance(colorscale, (str, type(None))):
+        colorscale = px.colors.get_colorscale(colorscale or "viridis")
+    elif not isinstance(colorscale, Sequence) or not isinstance(
+        colorscale[0], (str, list, tuple)
+    ):
+        raise TypeError(
+            f"{colorscale=} should be string, list of strings or list of "
+            "tuples(float, str)"
+        )
+
+    color_bar = color_bar or {}
+    color_bar.setdefault("orientation", "h")
+    # if values is a series with a name, use it as the colorbar title
+    if isinstance(values, pd.Series) and values.name:
+        color_bar.setdefault("title", values.name)
+
+    values = count_elements(values, count_mode, exclude_elements, fill_value)
+
+    if heat_mode in ("fraction", "percent"):
+        # normalize heat values
+        clean_vals = values.replace([np.inf, -np.inf], np.nan).dropna()
+        # ignore inf values in sum() else all would be set to 0 by normalizing
+        heat_value_element_map = values / clean_vals.sum()
+    else:
+        heat_value_element_map = values
+
+    n_rows, n_columns = 10, 18
+    # initialize tile text and hover tooltips to empty strings
+    tile_texts, hover_texts = np.full([2, n_rows, n_columns], "", dtype=object)
+    heatmap_values = np.full([n_rows, n_columns], np.nan)
+
+    if label_map is None:
+        # default to space string for None, np.nan and "nan". space is needed
+        # for <br> in tile_text to work so all element symbols are vertically aligned
+        label_map = dict.fromkeys([np.nan, None, "nan"], " ")  # type: ignore[list-item]
+
+    for symbol, period, group, name, *_ in df_ptable.itertuples():
+        # build table from bottom up so that period 1 becomes top row
+        row = n_rows - period
+        col = group - 1
+
+        label = ""  # label (if not None) is placed below the element symbol
+        if show_values:
+            if symbol in exclude_elements:
+                label = "excl."
+            elif heat_value := heat_value_element_map.get(symbol):
+                if heat_mode == "percent":
+                    label = f"{heat_value:{fmt or '.1%'}}"
+                else:
+                    default_prec = ".1f" if heat_value < 100 else ",.0f"
+                    if heat_value > 1e5:
+                        default_prec = ".2g"
+                    label = f"{heat_value:{fmt or default_prec}}".replace("e+0", "e")
+
+            if callable(label_map):
+                label = label_map(label)
+            elif isinstance(label_map, dict):
+                label = label_map.get(label, label)
+        style = f"font-weight: bold; font-size: {1.5 * (font_size or 12)};"
+        tile_text = (
+            f"<span {style=}>{symbol}</span>{f'<br>{label}' if show_values else ''}"
+        )
+
+        tile_texts[row][col] = tile_text
+
+        hover_text = name
+
+        if hover_data is not None and symbol in hover_data:
+            hover_text += f"<br>{hover_data[symbol]}"
+
+        if hover_props is not None:
+            if unsupported_keys := set(hover_props) - set(df_ptable):
+                raise ValueError(
+                    f"Unsupported hover_props: {', '.join(unsupported_keys)}. Available"
+                    f" keys are: {', '.join(df_ptable)}.\nNote that some keys have "
+                    "missing values."
+                )
+            df_row = df_ptable.loc[symbol]
+            if isinstance(hover_props, dict):
+                for col_name, col_label in hover_props.items():
+                    hover_text += f"<br>{col_label} = {df_row[col_name]}"
+            elif isinstance(hover_props, (list, tuple)):
+                hover_text += "<br>" + "<br>".join(
+                    f"{col_name} = {df_row[col_name]}" for col_name in hover_props
+                )
+            else:
+                raise ValueError(
+                    f"hover_props must be dict or sequence of str, got {hover_props}"
+                )
+
+        hover_texts[row][col] = hover_text
+
+        # TODO maybe there's a more elegant way to handle excluded elements?
+        if symbol in exclude_elements:
+            continue
+
+        color_val = heat_value_element_map[symbol]
+        if log and color_val > 0:
+            color_val = np.log10(color_val)
+        # until https://github.com/plotly/plotly.js/issues/975 is resolved, we need to
+        # insert transparency (rgba0) at low end of colorscale (+1e-6) to not show any
+        # colors on empty tiles of the periodic table
+        heatmap_values[row][col] = color_val
+
+    if isinstance(font_colors, str):
+        font_colors = [font_colors]
+    if cscale_range == (None, None):
+        cscale_range = (values.min(), values.max())
+
+    fig = ff.create_annotated_heatmap(
+        heatmap_values,
+        annotation_text=tile_texts,
+        text=hover_texts,
+        showscale=show_scale,
+        colorscale=colorscale,
+        font_colors=font_colors or None,
+        hoverinfo="text",
+        xgap=gap,
+        ygap=gap,
+        zmin=None if log else cscale_range[0],
+        zmax=None if log else cscale_range[1],
+        # zauto=False if cscale_range is set, needed for zmin, zmax to work
+        # see https://github.com/plotly/plotly.py/issues/193
+        zauto=cscale_range == (None, None),
+        **kwargs,
+    )
+    fig.update_layout(
+        margin=dict(l=10, r=10, t=10, b=10, pad=10),
+        paper_bgcolor=bg_color,
+        plot_bgcolor="rgba(0, 0, 0, 0)",
+        xaxis=dict(zeroline=False, showgrid=False),
+        yaxis=dict(zeroline=False, showgrid=False, scaleanchor="x"),
+        font_size=font_size,
+        width=1000,
+        height=500,
+    )
+
+    if color_bar.get("orientation") == "h":
+        dct = dict(x=0.4, y=0.75, titleside="top", len=0.4)
+        color_bar = {**dct, **color_bar}
+    else:  # make title vertical
+        dct = dict(titleside="right", len=0.87)
+        color_bar = {**dct, **color_bar}
+        if title := color_bar.get("title"):
+            # <br><br> to increase title offset
+            color_bar["title"] = f"<br><br>{title}"
+
+    fig.update_traces(colorbar=dict(lenmode="fraction", thickness=15, **color_bar))
     return fig
