--- pymatviz/ptable/__init__.py
@@ -1,13 +1,0 @@
-"""matplotlib and plotly periodic table figures."""
-
-from __future__ import annotations
-
-from pymatviz.ptable.matplotlib import (
-    ptable_heatmap,
-    ptable_heatmap_ratio,
-    ptable_heatmap_splits,
-    ptable_hists,
-    ptable_lines,
-    ptable_scatters,
-)
-from pymatviz.ptable.plotly import ptable_heatmap_plotly
